rename and clean the files
