+++ src/features/MergeNetworks/models/Impl/mergeNetwork.tsx
@@ -1,12 +1,12 @@
-import { IdType } from "../../models/IdType";
-import TableFn from "../../models/TableModel";
-import { NetworkRecord } from "./models/DataInterfaceForMerge";
-import NetworkFn, { Edge, Network, Node } from "../../models/NetworkModel";
-import { Column } from "../../models/TableModel/Column";
-import { ValueType } from "../../models/TableModel/ValueType";
-import { attributeValueMatcher } from "./utils/attributesOperations";
-import { MatchingTable } from "./models/Impl/MatchingTable";
-import { preprocess, castAttributes, addMergedAtt } from "./utils/attributesOperations";
+import { IdType } from "../../../../models/IdType";
+import TableFn from "../../../../models/TableModel";
+import { NetworkRecord } from "../DataInterfaceForMerge";
+import NetworkFn, { Edge, Network, Node } from "../../../../models/NetworkModel";
+import { Column } from "../../../../models/TableModel/Column";
+import { ValueType } from "../../../../models/TableModel/ValueType";
+import { attributeValueMatcher } from "../../utils/attributes-operations";
+import { MatchingTable } from "./MatchingTable";
+import { preprocess, castAttributes, addMergedAtt } from "../../utils/attributes-operations";
 
 export function mergeNetwork(fromNetworks: IdType[], toNetworkId: IdType, networkRecords: Record<IdType, NetworkRecord>,
     nodeAttributeMapping: MatchingTable, edgeAttributeMapping: MatchingTable, networkAttributeMapping: MatchingTable, matchingAttribute: Record<IdType, Column>) {
