+++ src/features/MergeNetworks/utils/helper-functions.ts
